Adding Exercises

+++ Gammal_Tech Academy/Core_Courses/1_C_Programming_Language/Les_41_Enum/Examples/Example_2.c
@@ -22,7 +22,7 @@ int main()
         Dec
     };
 
-    for (size_t i = 0; i <= Dec; i++)
+    for (int i = 0; i <= Dec; i++)
     {
         printf("\n%d", i);
     }
