+++ Gammal_Tech Academy/Core_Courses/1_C_Programming_Language/Les_41_Enum/Examples/Example_3.c
@@ -22,7 +22,7 @@ int main()
         Dec
     };
 
-    for (size_t i = Jan; i <= Dec; i++)
+    for (int i = Jan; i <= Dec; i++)
     {
         printf("\n%d", i);
     }
